Conclusão da implementação do cache de antenas como o pré-carregamento.

+++ app/src/main/java/devandroid/adenilton/erbazimuth/data/model/CellTowerInfo.kt
@@ -4,8 +4,8 @@ package devandroid.adenilton.erbazimuth.data.model
 data class CellTowerInfo(
     val mcc: Int, // Mobile Country Code
     val mnc: Int, // Mobile Network Code
-    val lac: Int, // Location Area Code
-    val cid: Int, // Cell ID
+    val lac: Long, // Location Area Code
+    val cid: Long, // Cell ID
     val signalStrength: Int, // Intensidade do Sinal (dBm)
     val operatorName: String // Nome da operadora (Ex: "TIM", "Claro")
 )
